Move util method to correct Class
