+++ src/main/java/ch/heigvd/pro/b04/sessions/SessionIdentifier.java
@@ -7,13 +7,32 @@ import javax.persistence.Column;
 import javax.persistence.Embeddable;
 import javax.persistence.ManyToOne;
 import javax.persistence.PrimaryKeyJoinColumn;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 @Embeddable
 @NoArgsConstructor
+@AllArgsConstructor
+@Builder
 public class SessionIdentifier implements Serializable {
+
+  /**
+   * Creates a new unique identifier for a Session.
+   * @param repository The repository to which we will add a new Session to
+   * @return A new unique identifier
+   */
+  public static Long getNewIdentifier(SessionRepository repository) {
+    Long identifier = repository.findAll().stream()
+        .map(ServerSession::getIdSession)
+        .map(SessionIdentifier::getIdSession)
+        .max(Long::compareTo)
+        .map(id -> id + 1)
+        .orElse(1L);
+    return identifier;
+  }
 
   @Getter
   @Column
@@ -27,8 +46,4 @@ public class SessionIdentifier implements Serializable {
   @ManyToOne
   @PrimaryKeyJoinColumn
   private ServerPoll idxPoll;
-
-  public SessionIdentifier(long id) {
-    this.idSession = id;
-  }
 }
